backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.7
SQLAlchemy==2.0.36
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
email-validator==2.2.0
pytest==8.3.3
